Remove official 3D model library submodule (#20)
Resolves #19 (Remove official 3D model library submodule).

The set of symbol libraries and footprint libraries available for use by
a project are managed using tables. This is not done for 3D model
libraries. Instead, footprints contain the path to the associated 3D
model relative to a KiCad environment variable (`KISYS3DMOD` for the
official libraries). The set of standard KiCad environment variables
cannot support a project structure where libraries are managed in a
single Git repository that is then used as a submodule by projects.

Using a smudge filter (as suggested by
https://bugs.launchpad.net/kicad/+bug/1840458/comments/4) and/or placing
the 3D models in their own repository could be used as potential
workarounds for this sometime in the future.